refactor: Store 모듈 분리 (#33)
