+++ src/main/java/goorm/_44/controller/file/FileController.java
@@ -3,7 +3,7 @@ package goorm._44.controller.file;
 import goorm._44.config.api.ApiResult;
 import goorm._44.dto.request.PresignRequest;
 import goorm._44.dto.response.PresignResponse;
-import goorm._44.service.user.PresignService;
+import goorm._44.service.file.PresignService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
