+++ src/main/java/goorm/_44/dto/response/IdResponse.java
@@ -1,0 +1,3 @@
+package goorm._44.dto.response;
+
+public record IdResponse(Long id) {}
